requirements.txt
# === Bot 核心依赖 ===
python-telegram-bot==21.0.1
httpx~=0.27.0  # 修复版本冲突（与 python-telegram-bot 兼容）

# === 数据库 ===
sqlalchemy>=2.0.0
alembic==1.13.0  # 数据库迁移
pymysql>=1.1.0  # MySQL 驱动（Zeabur 部署需要）

# === Web 框架 ===
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# === 存储 ===
redis>=5.0.0

# === 异步任务队列（Stage 3）===
arq==0.25.0
tenacity==8.2.3
APScheduler==3.10.4  # 定时任务调度器（用于订单超时处理）

# === 可观测性（Stage 4）===
structlog==24.1.0
loguru==0.7.2  # 日志库（用于 energy 模块）
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0

# === 中间件（Stage 5）===
slowapi==0.1.9  # Rate limiting
pybreaker==1.0.2  # Circuit breaker
python-multipart==0.0.6  # For form data parsing

# === Streamlit 管理界面（Stage 7）===
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
python-dotenv==1.0.0

# === 测试 ===
pytest==7.4.3
pytest-asyncio>=0.23.0
pytest-timeout>=2.3.0
pytest-cov>=4.1.0
fakeredis>=2.20.0